Replace units float with :decimal
- modify the create_wallets migration
- modify seeds
- modify factory
- modify Wallet schema
- replace ell the definitions of units :float with :decimal and specify the default value if needed
- create TestHelpers module with units_to_string/1 function
- update tests

+++ lib/payment_server_web/graphql/mutations/wallet_mutation.ex
@@ -9,7 +9,8 @@ defmodule PaymentServerWeb.Graphql.Mutations.WalletMutation do
     field :create_wallet, :wallet do
       arg(:currency, non_null(:currency_type))
       arg(:user_id, non_null(:id))
-      arg(:units, :float)
+      # , default_value: 0.0)
+      arg(:units, :decimal)
 
       resolve(&Resolvers.WalletResolver.create_wallet/3)
     end
@@ -18,7 +19,7 @@ defmodule PaymentServerWeb.Graphql.Mutations.WalletMutation do
     field :send_money, :sent_money do
       arg(:sender_wallet_id, non_null(:id))
       arg(:receiver_wallet_id, non_null(:id))
-      arg(:amount, non_null(:float))
+      arg(:amount, non_null(:decimal))
 
       resolve(&Resolvers.WalletResolver.send_money/3)
     end
